+++ src/Htdp/Shape.hs
@@ -26,65 +26,65 @@ solid = Solid
 outline = Outline
 
 applyMode :: Mode -> Color -> Image -> Image
-applyMode Solid color image = image # D.fc color
+applyMode Solid color image = image # D.fcA color
                                     # D.lcA transparent
-applyMode Outline color image = image # D.lc color
+applyMode Outline color image = image # D.lcA color
 
-circle :: Float -> Mode -> Color -> Image
+circle :: Double -> Mode -> Color -> Image
 circle r mode color = D.circle r # applyMode mode color
 
-ellipse :: Float -> Float -> Mode -> Color -> Image
+ellipse :: Double -> Double -> Mode -> Color -> Image
 ellipse w h mode color = D.ellipseXY (w / 2) (h / 2) # applyMode mode color
 
-line :: Float -> Float -> Color -> Image
-line x y color = D.strokeLine (D.fromVertices [0 ^& 0, x ^& (-y)]) # D.lc color
+line :: Double -> Double -> Color -> Image
+line x y color = D.strokeLine (D.fromVertices [0 ^& 0, x ^& (-y)]) # D.lcA color
 
-addLine :: Image -> Float -> Float -> Float -> Float -> Color -> Image
+addLine :: Image -> Double -> Double -> Double -> Double -> Color -> Image
 addLine image x0 y0 x1 y1 color =
-  (D.strokeLocLine (D.fromVertices [x0 ^& (-y0), x1 ^& (-y1)]) # D.lc color)
+  (D.strokeLocLine (D.fromVertices [x0 ^& (-y0), x1 ^& (-y1)]) # D.lcA color)
   `D.atop`
   (image # D.alignL # D.alignT)
 
 emptyImage :: Image
 emptyImage = mempty
 
-triangle :: Float -> Mode -> Color -> Image
+triangle :: Double -> Mode -> Color -> Image
 triangle sideLength mode color = D.triangle sideLength # applyMode mode color
 
--- rightTriangle :: Float -> Float -> Mode -> Color -> Image
+-- rightTriangle :: Double -> Double -> Mode -> Color -> Image
 -- rightTriangle = _
 
--- isoscelesTriangle :: Float -> Float -> Mode -> Color -> Image
+-- isoscelesTriangle :: Double -> Double -> Mode -> Color -> Image
 -- isoscelesTriangle = _
 
--- triangleSSS :: Float -> Float -> Float -> Mode -> Color -> Image
+-- triangleSSS :: Double -> Double -> Double -> Mode -> Color -> Image
 -- triangleSSS = _
 
--- triangleASS :: Float -> Float -> Float -> Mode -> Color -> Image
+-- triangleASS :: Double -> Double -> Double -> Mode -> Color -> Image
 -- triangleASS = _
 
--- triangleSAS :: Float -> Float -> Float -> Mode -> Color -> Image
+-- triangleSAS :: Double -> Double -> Double -> Mode -> Color -> Image
 -- triangleSAS = _
 
--- triangleSSA :: Float -> Float -> Float -> Mode -> Color -> Image
+-- triangleSSA :: Double -> Double -> Double -> Mode -> Color -> Image
 -- triangleSSA = _
 
--- triangleAAS :: Float -> Float -> Float -> Mode -> Color -> Image
+-- triangleAAS :: Double -> Double -> Double -> Mode -> Color -> Image
 -- triangleAAS = _
 
--- triangleASA :: Float -> Float -> Float -> Mode -> Color -> Image
+-- triangleASA :: Double -> Double -> Double -> Mode -> Color -> Image
 -- triangleASA = _
 
--- triangleSAA :: Float -> Float -> Float -> Mode -> Color -> Image
+-- triangleSAA :: Double -> Double -> Double -> Mode -> Color -> Image
 -- triangleSAA = _
 
-square :: Float -> Mode -> Color -> Image
+square :: Double -> Mode -> Color -> Image
 square size mode color = D.square size # applyMode mode color
 
-rectangle :: Float -> Float -> Mode -> Color -> Image
+rectangle :: Double -> Double -> Mode -> Color -> Image
 rectangle w h mode color = D.rect w h # applyMode mode color
 
-rhombus :: Float -> Float -> Mode -> Color -> Image
+rhombus :: Double -> Double -> Mode -> Color -> Image
 rhombus sideLength angle mode color = D.polygon polyOpts # applyMode mode color
   where
     verticalRadius = sideLength * D.cosA ((angle / 2) @@ D.deg)
@@ -94,5 +94,5 @@ rhombus sideLength angle mode color = D.polygon polyOpts # applyMode mode color
       D.NoOrient
       (D.P $ D.V2 0 0)
 
--- star :: Float -> Mode -> Color -> Image
+-- star :: Double -> Mode -> Color -> Image
 -- star = _
